+++ util/math_utils.c
@@ -2,24 +2,24 @@
 #include <math.h>
 #include <stdlib.h>
 
-double generate_std_norm_dist() {
-	double irwin_hall_var = 0.0;
+float generate_std_norm_dist() {
+	float irwin_hall_var = 0.0;
 	for (int i = 0; i < 12; i++) {
-		irwin_hall_var += (double)rand() / (double)RAND_MAX;
+		irwin_hall_var += (float)rand() / (float)RAND_MAX;
 	}
 	return min(1,max(-1,irwin_hall_var - 6));
 }
 
-double sigmoid(const double z) {
+float sigmoid(const float z) {
 	return 1 / (1 + exp(-z));
 }
 
-double sigmoid_prime(const double z) {
-	double ez = exp(-z);
+float sigmoid_prime(const float z) {
+	float ez = exp(-z);
 	return ez / ((1 + ez) * (1 + ez));
 }
 
-double tanh_prime(const double z) {
+float tanh_prime(const float z) {
 	return (1. - tanh(z) * tanh(z));
 }
