+++ util/math_utils.h
@@ -1,13 +1,13 @@
 #ifndef MATH_UTILS_H
 #define MATH_UTILS_H
 
-double generate_std_norm_dist();
+float generate_std_norm_dist();
 
-double sigmoid(const double z);
+float sigmoid(const float z);
 
-double sigmoid_prime(const double z);
+float sigmoid_prime(const float z);
 
-double tanh_prime(const double z);
+float tanh_prime(const float z);
 
 int byte_array_to_big_endian(unsigned char* bytes);
